+++ app/Http/Controllers/Api/AddressController.php
@@ -5,18 +5,23 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreAddressRequest;
 use App\Http\Requests\UpdateAddressRequest;
-use App\Models\Employee;
 use App\Models\Address;
+use App\Models\Employee;
+use Illuminate\Support\Facades\Gate;
 
 class AddressController extends Controller
 {
     public function index(Employee $employee)
     {
+        Gate::authorize('view', $employee);
+
         return $employee->addresses;
     }
 
     public function store(StoreAddressRequest $request, Employee $employee)
     {
+        Gate::authorize('update', $employee);
+
         $addresses = [];
 
         foreach ($request->validated()['addresses'] as $addressData) {
@@ -29,13 +34,17 @@ class AddressController extends Controller
 
     public function show(Employee $employee, Address $address)
     {
+        Gate::authorize('view', $employee);
+
         return $address;
     }
 
     public function update(UpdateAddressRequest $request, $employee_id)
     {
         $employee = Employee::where('id', $employee_id)->first();
-        
+
+        Gate::authorize('update', $employee);
+
         $employee->addresses()->delete();
 
         $addresses = [];
@@ -50,7 +59,10 @@ class AddressController extends Controller
 
     public function destroy(Employee $employee, Address $address)
     {
+        Gate::authorize('update', $employee);
+
         $address->delete();
+
         return response()->json(null, 204);
     }
 }
